fix: compatibility issues of dev server in iOS 10
