+++ .changeset/quiet-horses-rhyme.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/prod-server': patch
+'@modern-js/utils': patch
+---
+
+fix: compatibility issues of dev server in iOS 10
+
+fix: 修复 dev server 代码在 iOS 10 下的兼容性问题
